Closes #75 Add buttons corresponding to menu items
Adding buttons to the list-item is a bit tricky, as those list-items are
in fact xbl-bindings. Hence we create our own binding we use to rebind the
controls-container the list-item binding contains.

Adding buttons to the detailed view is pretty straight-forward; just an
overlay.

The commands and string-setup was refactored and split into a broadcaster and command each to account for the new binding and new overlay buttons.
(cherry picked from commit cf8b6e9087568f528973c9635f4b996a87ca327f)