Using broadcasters to update attribute values of elements in the toolbar menu-button instead of listening to targets to find out when the toobar menu-button is added to the browser.
The upside is that we don't have to figure out what all of the targets are and we don't have to listen to their DOMNodeInsterted events. The downside is that menu commands are not added to the menu immediately after the toolbar menu-button is added to the browser, instead they are added when the location is changed or when the menu is opened, after the menu-button is added to the browser.