Newscript: Validate that name and id are provided
If no name or id are given, accepting the dialog will fail and the
corresponding fields will be highlighted.
Also, the required fields got a placeholder text now, asking the user to
specifiy the text.

(Note: Not using HTML5 Form validation here, as this does not work
reliably when intermixed with XUL)